+++ packages/demo/src/plugins.tsx
@@ -6,11 +6,11 @@ import { geogebraPlugin } from '@edtr-io/plugin-geogebra'
 import { h5pPlugin } from '@edtr-io/plugin-h5p'
 import { highlightPlugin } from '@edtr-io/plugin-highlight'
 import { hintPlugin } from '@edtr-io/plugin-hint'
-import { createImagePlugin, ImageUploadConfig } from '@edtr-io/plugin-image'
+import { imagePlugin, readFile } from './plugin-image'
 import {
   createFilePlugin,
-  FileUploadConfig,
-  parseFileType
+  parseFileType,
+  UploadedFile
 } from '@edtr-io/plugin-files'
 import { inputExercisePlugin } from '@edtr-io/plugin-input-exercise'
 import { textPlugin } from '@edtr-io/plugin-text'
@@ -20,55 +20,27 @@ import { solutionPlugin } from '@edtr-io/plugin-solution'
 import { spoilerPlugin } from '@edtr-io/plugin-spoiler'
 import { videoPlugin } from '@edtr-io/plugin-video'
 
-interface SerloResponse {
-  files: { location: string; filename: string }[]
-}
-
-const baseUploadConfig = {
-  url: 'https://de.serlo.org/attachment/upload',
-  paramName: 'attachment[file]',
-  maxFileSize: 2 * 1024 * 1024,
-  getAdditionalFields: () => {
+const mockUploadFileHandler = (file: File): Promise<UploadedFile> => {
+  return readFile(file).then(loaded => {
     return {
-      type: 'file',
-      csrf: ((window as unknown) as { csrf: string }).csrf
+      location: loaded.dataUrl,
+      name: loaded.file.name,
+      type: parseFileType(loaded.file.name)
     }
-  }
-}
-export const imageUploadConfig: ImageUploadConfig<SerloResponse> = {
-  ...baseUploadConfig,
-  allowedExtensions: ['gif', 'jpg', 'jpeg', 'png', 'svg'],
-  getStateFromResponse: response => {
-    return {
-      src: response.files[0].location
-    }
-  }
-}
-
-export const fileUploadConfig: FileUploadConfig<SerloResponse> = {
-  ...baseUploadConfig,
-  getStateFromResponse: response => {
-    return {
-      location: response.files[0].location,
-      name: response.files[0].filename,
-      type: parseFileType(response.files[0].filename)
-    }
-  }
+  })
 }
 
 export const plugins: Record<string, Plugin> = {
+  // Must be placed before files for onPaste
+  image: imagePlugin,
   anchor: anchorPlugin,
   blockquote: blockquotePlugin,
   equations: equationsPlugin,
-  files: createFilePlugin({ upload: fileUploadConfig }),
+  files: createFilePlugin({ upload: mockUploadFileHandler }),
   h5p: h5pPlugin,
   geogebra: geogebraPlugin,
   highlight: highlightPlugin,
   hint: hintPlugin,
-  image: createImagePlugin({
-    upload: imageUploadConfig,
-    secondInput: 'description'
-  }),
   inputExercise: inputExercisePlugin,
   rows: rowsPlugin,
   scMcExercise: scMcExercisePlugin,