+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased](https://github.com/edtr-io/edtr-io/compare/v0.6.0..HEAD)
+
+### Breaking Changes
+
+- **plugin-image**. Changed configs for createImagePlugin and removed Upload export
+
+### Added
+
+- **core**. Added StateType upload for file uploading.
+
 ## [0.6.0](https://github.com/edtr-io/edtr-io/compare/v0.5.0..v0.6.0)
 
 ### Breaking Changes